dbnet

+++ utils/post_processing/seg_detector_representer.py
@@ -141,7 +141,7 @@ class SegDetectorRepresenter():
         return expanded
 
     def get_mini_boxes(self, contour):
-        bounding_box = cv2.minAreaRect(contour)
+        bounding_box = cv2.minAreaRect(np.float32(contour))
         points = sorted(list(cv2.boxPoints(bounding_box)), key=lambda x: x[0])
 
         index_1, index_2, index_3, index_4 = 0, 1, 2, 3
